+++ src/wavewindow.h
@@ -22,16 +22,10 @@ void make_sigarea_gcs(GtkWidget *widget);
 void force_screengrab_gcs(void);
 void force_normal_gcs(void);
 
-void XXX_gdk_draw_line(cairo_t *cr, wave_rgb_t gc, gint _x1, gint _y1, gint _x2, gint _y2);
-void XXX_gdk_draw_rectangle(cairo_t *cr,
-                            wave_rgb_t gc,
-                            gboolean filled,
-                            gint _x1,
-                            gint _y1,
-                            gint _w,
-                            gint _h);
+void XXX_gdk_draw_line(cairo_t *cr, wave_rgb_t gc, gdouble _x1, gdouble _y1, gdouble _x2, gdouble _y2);
+void XXX_gdk_draw_rectangle(cairo_t *cr, wave_rgb_t gc, gboolean filled, gdouble _x1, gdouble _y1, gdouble _w, gdouble _h);
 void XXX_gdk_set_color(cairo_t *cr, wave_rgb_t gc);
-void XXX_gdk_draw_line2(cairo_t *cr, gint _x1, gint _y1, gint _x2, gint _y2);
+void XXX_gdk_draw_line2(cairo_t *cr, gdouble _x1, gdouble _y1, gdouble _x2, gdouble _y2);
 
 #ifdef WAVE_GTK3_SIZE_ALLOCATE_WORKAROUND_WAVE_VSLIDER
 void service_vslider(GtkWidget *text, gpointer data);
